+++ tests/TestCase.php
@@ -2,8 +2,22 @@
 
 namespace SebastiaanLuca\Module\Tests;
 
-use PHPUnit\Framework\TestCase as BaseTestCase;
+use Orchestra\Testbench\TestCase as BaseTestCase;
+use SebastiaanLuca\Module\ModuleServiceProvider;
 
 class TestCase extends BaseTestCase
 {
+    /**
+     * Get package providers.
+     *
+     * @param  \Illuminate\Foundation\Application $app
+     *
+     * @return array
+     */
+    protected function getPackageProviders($app) : array
+    {
+        return [
+            ModuleServiceProvider::class,
+        ];
+    }
 }