+++ CHANGELOG.md
@@ -6,7 +6,18 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## Unreleased
 
+## 0.1.0 (2018-07-25)
+
 ### Added
 
-- Module service provider
-- Automatically load all modules from a directory
+- Added module service provider
+- Added module autoloading
+- Added configurable scan paths
+- Added automated router mapping in the module service provider
+- Added configuration loading
+- Added translations loader
+- Added views loader
+- Added factories loader
+- Added migrations loader
+- Added seeders loader
+- Make configuration publishable